Extract CEHR page into per-concern components and a calculator hook
Splits the 643-line app/cehr/page.tsx into focused modules to make the
PR easier to review and the page easier to evolve:

- _Header.tsx — intro title + 3 explanatory paragraphs
- _EligibilityInputs.tsx — situation toggle + 3 RFR fields
- _RfrField.tsx — single input atom with status icon and tooltip
- _CehrComparisonTable.tsx — side-by-side table and expandable steps
- _AssetTypeSelector.tsx — asset checkboxes for the email template
- _EmailPreview.tsx — readonly email with copy-to-clipboard
- _References.tsx — Légifrance, BOFiP, prosper-conseil links
- _useCehrCalculator.ts — derived state (statuses, result, email)

page.tsx is now a 96-line orchestration that owns the form state and
wires the hook to the components.

Also addresses two Codex correctness findings:
- Round the CEHR total to the nearest euro per BOFiP, so the displayed
  totals are consistent with the computed savings.
- Centralize the fmtEur helper in lib/taxes/cehr.ts so it is shared by
  the page and the email template instead of being duplicated.
